Added support for reading in wedge elements
This was pretty straight forward. I basically just added in some code that identified the element type (using the Contains function).

I then modified the scaning part of the code so that it behaves differently when reading in a Wedge element.

My code runs just fine with these changes on the cylinder file (which does not have any wedge elements). I need a different inp file (one with wedge elements) to see if it works with those as well.

In my next commit, I plan on adding write support for wedge elements as well.