Rewrote how node sets are read in
I changed the part of the code that deals with reading in node sets. I realized that I really needed a way to read in a specific node set (if the file happens to have multiple node sets), and possibly apply different BC's to each one.

As such, I moved the node set part of IO::Read::inp into a new function, IO::Read::node_set. This function uses the same file opening process as IO::Read::inp, but only reads in lines that begin with *Nset.

The new function accepts an optional "Node_Set_Name" argument (which specifies the name of the node set that you want to read in). The code is pretty much the same otherwise.

I also realized that node sets come in two forms, "genrate" and "list". My code only deals with "generate" type node sets, but it needs to be able to deal with list type ones as well. I added in a placeholder for doing this, but haven't gotten to it (I think it's going to involve splitting strings, which sounds tricky in c++, oh boy).

The Node_Set_Name has a default value (just an end of string character), though I'm not sure if it's working.

The code appears to run, but I don't know if the new function works as intended. I need to test that and write the part that deals with list type node sets. Expect that tomorrow (hopefully).