I wrote a split function
To deal with the CSV format used in "list" type node sets, I had to write some kind of split function for strings. I tried finding an existing function, but they all seemed recondite... so I wrote my own!

The new function, IO::Read::Split, basically works as follows: It keeps track of two indicies, Index_Start and N_Chars_Since_Delim. Index_Start is simply the index of the first character after the last instance of delim. N_Chars_Since_Delim is the number of characters after Index_Start that are not the delimeter. The code cycles through the characters in the passed string. If the current character matches the delimeter, then a new substring is crated using the N_Chars_Since_Delim characters in S starting at Index_Start. Index_Start and N_Chars_Since_Delim are then updated. If there is no match, then N_Chars_Since_Delim is incremented.

There is a special "clean up" part located after the loop to deal with the characters in S after the last instance of delim.

I also wrote some tests for this code, and it appears to work!

In my next update, I plan on moving this function and Contains into a special String file (likely in the IO directory). It doesn't make sense for these functions, which can be used in a lot more than just the inp reader, to be isolated in IO::Read (and the inp_reader directory).