Rewrote Split to work with char arrays + (maybe) added list reading for node sets
I realized that my inp reader works using char arrays, not strings. This was a problem, since my split function expects string inputs. I thought about casting buffer to a string and then passing it to the string-based reader, but concluded that this would be very inefficient. Thus, in the interest of performance, I decided to instead rewrite the Split method to work with char arrays. I wrote a new version of Split that accepts char arrays (the string based one just calls this function).

Since the new version uses char arrays instead of string objects, I had to find ways to hard-code substr. I also had to figure out how to determine the char array's length without manually counting it (it's just the value of i after going through the main loop). The for loop was replaced with a while loop (it goes until it reaches a '\0'.

The code still returns a vector of strings (which, admittedly, is a little odd given that the function accepts c strings as it's input... whatever, it does what I need it to do right now).

The code appears to work with these changes.

I also finished (I think) the "make nset from list"  part of the Node set reader (See inp_reader.cc). I have not tested this code, however.

In my next commit, I plan on testing the node set reader. If I can get it working, I will attempt to run the simulation sent to me by murdang.