Got node set working + improved split
I got the node set reader working (with list type node sets).

In the process, I had to fix several bugs in my code. For one, there was an infinite loop near the start of the node_set reader. If a node set's name didn't match the specified name, the code simply continued. The issue is that it didn't read in a new line, so it just got stuck in an infinite loop. A simple 1 line change fixed that (see inp_reader.cc:228).

I also fixed the breakout conditions in the inp reader (which was using strcmp... oh god). See inp_reader.cc:267 and 287.

I also put in a check to skip empty sets (for list reading). See inp_reader.cc:297.

I also had to make some tweaks to the split method. In particular, I needed to add checks for new line and carage return characters (when deciding if I should break out). See String_Ops.cc:84.

I also decided to change the index for the inner while loops from i to j to avoid confusion (with the other variable i).

I then modified the simulation code so that it reads in a particular node set and checks that the read in data is what I expect it to be. Based on this, I think it's working.

In my next commit, I want to make a new test (probably like "Simulation Test" or something) to test Murdang's stuff. At that time, I'll restore simulation.cc to what it was before this commit.