Made simulation test + overhauled how nset BC's are applied
The changes that I made to simulation in the last commit have been moved to a new file called Simulation_Tests. I haven't done anything with these tests yet.

More importantly, I reworte the way that BC's are applied to nodes in a node set. Before, I had it set up such that every node in a node set had its BC's all fixed. This won't work with Mrudang's test. I had to rewrite the way that this works.

To begin, I made a new function called Set_nset_BC in Simulation.h. I got rid of the Node_Set_List parameter for Process_Node_List (and removed that part of the function). The new function works by accepting the Node_Set_List in addition to a "nset_BC" object. This object basically stores information on how to apply BC's to the nod set. The class definition is in inp_reader.h,  but it's pretty simple (baiscally just a fancy structure). Each object stores which components have BC's, as well as what those BC's are. When you want to set the BC's in a node set to some value, you first set up one of these nset_BC objects. Once you've done that, the Node_Set_List along with the nset_BC object is passed to Set_nset_BC. From ther, the BC's are set according to data in the nset_BC object. See Simulation.cc:200 for implementation.

With this, I got the old cylinder simulation running again. It appears to work as indended (with the new system for setting node set BCs!)

In the next commit, I plan on writing all of Mrudang's test (in Simulation_Tests.cc) to test the validity of my code.