Fixed a bug in Job-1.inp + added Fill method (Matrix)
I added a fill method for the matrix class. This replaced the existing "Zero" method. It does exactly what "Zero" did, but you can not speicify the value that you want to fill the matrix with.

I made corresponding changes elsewhere in the code.

I also found a bug in Job-1.inp. The 6th node set (containing nodes 17, 18, and 188) was missing. I fixed that.